requirements.txt
# Dev/Deployment
sphinx
sphinx_rtd_theme
nose
coverage
pypi-publisher
hug
googlemaps
PyJWT
marshmallow
python-dateutil
marshmallow-sqlalchemy
SQLAlchemy